requirements.txt
Flask-RESTful
Flask-JWT
